refactor(e2e): promote storage case pure helpers to shared module
storage_v2_upgrade needs the same history-recording/anchor-spec/offline-env
logic that storage_v2_baseline kept case-local, so move
storage_baseline_lib.py to gravity_e2e/helpers/storage_case_lib.py
(BaselineHistory renamed OnChainHistory; logic unchanged). TC1 now imports
the shared module and the shared node_process helper; its unit tests split
into test_storage_case_lib.py (moved lib tests) and the case-local
render/consistency tests that remain in test_storage_baseline_case.py.
