+++ gravity_e2e/gravity_e2e/helpers/storage_case_lib.py
@@ -1,12 +1,12 @@
 """
-Pure helpers for the storage_v2_baseline case.
+Pure helpers shared by the storage-v2 cluster cases
+(storage_v2_baseline / storage_v2_upgrade).
 
 Everything here is deterministic and unit-tested in
-gravity_e2e/tests/unit/test_storage_baseline_case.py (loaded by file path;
-this directory is not a package). The pytest case keeps orchestration and
-assertions; this module keeps the derivable facts:
+gravity_e2e/tests/unit/test_storage_case_lib.py. The pytest cases keep
+orchestration and assertions; this module keeps the derivable facts:
 
-- BaselineHistory / build_anchor_spec: turn the tx receipts recorded while
+- OnChainHistory / build_anchor_spec: turn the tx receipts recorded while
   building history into the H1 AnchorSpec, covering all six anchor kinds
   (balance, storage, transaction, receipt, logs, block_hash).
 - derive_offline_env: encode the on-disk layout that cluster/deploy.sh
@@ -14,6 +14,10 @@ assertions; this module keeps the derivable facts:
   what the node ran with.
 - encode_set_call: calldata for AnchorTarget.set(uint256) without going
   through web3 contract ABI codecs (stable across web3 v6/v7).
+
+History: extracted verbatim from storage_v2_baseline's case-local
+storage_baseline_lib.py when storage_v2_upgrade needed the same logic;
+``BaselineHistory`` kept as an alias for the recorded name.
 """
 
 from __future__ import annotations
@@ -42,8 +46,8 @@ class TxPoint:
 
 
 @dataclass(frozen=True)
-class BaselineHistory:
-    """What the case did on-chain, as recorded from receipts.
+class OnChainHistory:
+    """What a case did on-chain, as recorded from receipts.
 
     Attributes:
         faucet: sender address funding the transfers.
@@ -64,6 +68,10 @@ class BaselineHistory:
     sets: List[TxPoint]
 
 
+# Name under which storage_v2_baseline originally recorded this dataclass.
+BaselineHistory = OnChainHistory
+
+
 def encode_set_call(value: int) -> str:
     """Calldata for AnchorTarget.set(value) as a 0x-hex string."""
     if not 0 <= value < 2**256:
@@ -71,14 +79,14 @@ def encode_set_call(value: int) -> str:
     return "0x" + SET_SELECTOR + format(value, "064x")
 
 
-def build_anchor_spec(history: BaselineHistory) -> AnchorSpec:
+def build_anchor_spec(history: OnChainHistory) -> AnchorSpec:
     """Build the H1 collection spec from the recorded history.
 
     Covers all six anchor kinds:
     - balance: recipient at every transfer block (different value each
       time) plus the faucet at the last transfer block;
     - storage: contract slot 0 at every set() block (different value each
-      time) — the changeset-backed historical reads this case exists for;
+      time) — the changeset-backed historical reads these cases exist for;
     - transaction + receipt: every recorded tx hash;
     - block_hash: every block touched by the history;
     - logs: one address-filtered range spanning the contract's lifetime
@@ -127,7 +135,8 @@ def derive_offline_env(
     cluster/templates/reth_config.json.tpl):
     - node dir: <base_dir>/<node_id> with bin/ config/ data/ logs/ script/;
     - node binary hardlinked to <node dir>/bin/gravity_node — using it here
-      guarantees the offline commands run the exact binary the node ran;
+      guarantees the offline commands run the exact binary the node ran
+      (in storage_v2_upgrade: the post-upgrade binary after the swap);
     - STORAGE_DIR = <node dir>/data, reth datadir = ${STORAGE_DIR}/reth;
     - --datadir.static-files is ALSO ${STORAGE_DIR}/reth (the datadir root,
       not the reth default <datadir>/static_files), so it must be passed
